+++ demisto_sdk/commands/common/hook_validations/playbook.py
@@ -40,6 +40,7 @@ class PlaybookValidator(ContentEntityValidator):
                 self.is_script_id_valid(id_set_file),
                 self._is_id_uuid(),
                 self._is_taskid_equals_id(),
+                self.does_name_contains_the_type()
             ]
             answers = all(new_playbook_checks)
         else:
@@ -56,6 +57,7 @@ class PlaybookValidator(ContentEntityValidator):
                 self.is_script_id_valid(id_set_file),
                 self._is_id_uuid(),
                 self._is_taskid_equals_id(),
+                self.does_name_contains_the_type()
             ]
             answers = all(modified_playbook_checks)
 
@@ -401,3 +403,18 @@ class PlaybookValidator(ContentEntityValidator):
                 # invalid task in order to raise error for all the invalid tasks at the file
 
         return is_valid
+
+    def does_name_contains_the_type(self):
+        """
+        Check that the entity name does not contain the entity type
+        Returns: True if the name is valid
+        """
+
+        name = self.current_file.get('name', '')
+        if 'playbook' in name.lower():
+            error_message, error_code = Errors.entity_name_contains_type('name')
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                self.is_valid = False
+                return False
+        return True
+