+++ demisto_sdk/commands/common/tests/integration_test.py
@@ -793,6 +793,56 @@ class TestIntegrationValidator:
 
         assert not validator.check_separators_in_files()
 
+    def test_name_contains_the_type(self, pack):
+        """
+        Given
+            - An integration with a name that contains the type.
+        When
+            - running does_name_contains_the_type.
+        Then
+            - Ensure the validate failed.
+        """
+
+        integration = pack.create_integration(yml={"name": "test_integration"})
+
+        structure_validator = StructureValidator(integration.yml.path)
+        validator = IntegrationValidator(structure_validator)
+
+        assert not validator.does_name_contains_the_type()
+
+    def test_display_name_contains_the_type(self, pack):
+        """
+        Given
+            - An integration with a display name that contains the type.
+        When
+            - running does_name_contains_the_type.
+        Then
+            - Ensure the validate failed.
+        """
+
+        integration = pack.create_integration(yml={"display": "test_integration"})
+
+        structure_validator = StructureValidator(integration.yml.path)
+        validator = IntegrationValidator(structure_validator)
+
+        assert not validator.does_name_contains_the_type()
+
+    def test_name_does_not_contains_the_type(self, pack):
+        """
+        Given
+            - An integration with a name that does not contains the type.
+        When
+            - running does_name_contain_the_type.
+        Then
+            - Ensure the validate passes.
+        """
+
+        integration = pack.create_integration(yml={"name": "test", "display": "test"})
+
+        structure_validator = StructureValidator(integration.yml.path)
+        validator = IntegrationValidator(structure_validator)
+
+        assert validator.does_name_contains_the_type()
 
 class TestIsFetchParamsExist:
     def setup(self):